requirements.txt
# 测试框架
pytest==7.4.0
pytest-html==4.0.2
pytest-xdist==3.3.1
pytest-ordering==0.6
pytest-rerunfailures==12.0

# 报告
allure-pytest==2.13.0
allure-python-commons==2.13.0

# HTTP请求
requests==2.31.0

# 数据处理
jsonpath-ng==1.6.0
PyYAML==6.0

# 其他工具
colorlog==6.7.0